+++ src/main/java/com/asg/shipchandling/salesinvoice/dto/request/LoadQuotationItemsRequest.java
@@ -1,18 +1,17 @@
 package com.asg.shipchandling.salesinvoice.dto.request;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class LoadQuotationItemsRequest {
     private Long qtnPoid;
-    private Long incentiveAmt;
-    private Long incentiveAmt2;
-    private Long incentiveAmt3;
+    private BigDecimal incentiveAmt;
+    private BigDecimal incentiveAmt2;
+    private BigDecimal incentiveAmt3;
 }
